+++ nucleus/__init__.py
@@ -4,6 +4,7 @@ __all__ = [
     "AsyncJob",
     "EmbeddingsExportJob",
     "BoxAnnotation",
+    "DeduplicationJob",
     "DeduplicationResult",
     "DeduplicationStats",
     "BoxPrediction",
@@ -131,7 +132,11 @@ from .data_transfer_object.dataset_info import DatasetInfo
 from .data_transfer_object.job_status import JobInfoRequestPayload
 from .dataset import Dataset
 from .dataset_item import DatasetItem
-from .deduplication import DeduplicationResult, DeduplicationStats
+from .deduplication import (
+    DeduplicationJob,
+    DeduplicationResult,
+    DeduplicationStats,
+)
 from .deprecation_warning import deprecated
 from .errors import (
     DatasetItemRetrievalError,